+++ Backend/controllers/changepasswordcontroller.js
@@ -1,0 +1,53 @@
+const ChangePasswordController = require('../services/changePasswordService')
+const LoginService = require('../services/loginService')
+
+class ChangePasswordController {
+    constructor() {
+        this.loginService = new ChangePasswordService();
+    }
+
+    async firstLogin(req, res) {
+        try {
+          const { userId, newPassword, securityQuestions } = req.body;
+          console.log('Processing first login for user:', userId);
+          
+          const result = await this.loginService.updateFirstTimeUser(userId, newPassword, securityQuestions);
+          res.status(200).json({ message: 'Profile updated successfully', data: result });
+        } catch (error) {
+          console.error('Error in firstLogin:', error);
+          res.status(500).json({ 
+            message: 'Failed to update profile', 
+            error: error.message 
+          });
+        }
+      }
+
+      async changePassword(req, res) {
+        try {
+            const { userId, newPassword } = req.body;
+            const result = await this.loginService.changePassword(userId, newPassword);
+            res.status(200).json({ message: 'Password changed successfully', data: result });
+        } catch (error) {
+            res.status(500).json({ message: 'Internal server error', error: error.message });
+        }
+    }
+
+    
+    async firstLogin(req, res) {
+        try {
+          const { userId, newPassword, securityQuestions } = req.body;
+          console.log('Processing first login for user:', userId);
+          
+          const result = await this.loginService.updateFirstTimeUser(userId, newPassword, securityQuestions);
+          res.status(200).json({ message: 'Profile updated successfully', data: result });
+        } catch (error) {
+          console.error('Error in firstLogin:', error);
+          res.status(500).json({ 
+            message: 'Failed to update profile', 
+            error: error.message 
+          });
+        }
+      }
+}   
+
+module.exports = ChangePasswordController;
