+++ Backend/controllers/logincontroller.js
@@ -30,22 +30,6 @@ class LoginController {
         }
     }
 
-    async firstLogin(req, res) {
-        try {
-          const { userId, newPassword, securityQuestions } = req.body;
-          console.log('Processing first login for user:', userId);
-          
-          const result = await this.loginService.updateFirstTimeUser(userId, newPassword, securityQuestions);
-          res.status(200).json({ message: 'Profile updated successfully', data: result });
-        } catch (error) {
-          console.error('Error in firstLogin:', error);
-          res.status(500).json({ 
-            message: 'Failed to update profile', 
-            error: error.message 
-          });
-        }
-      }
-
     //For checking if the user is first time login or not
     async checkUserStatus(req, res) {
         try {
@@ -66,16 +50,6 @@ class LoginController {
                 message: 'Error checking user status', 
                 error: error.message 
             });
-        }
-    }
-
-    async changePassword(req, res) {
-        try {
-            const { userId, newPassword } = req.body;
-            const result = await this.loginService.changePassword(userId, newPassword);
-            res.status(200).json({ message: 'Password changed successfully', data: result });
-        } catch (error) {
-            res.status(500).json({ message: 'Internal server error', error: error.message });
         }
     }
