+++ Backend/routes/changepasswordroutes.js
@@ -3,11 +3,5 @@ const router = express.Router();
 const LoginController = require('../controllers/logincontroller'); // Import the controller
 const ChangePasswordController = require('../controllers/changepasswordcontroller');
 
-const loginController = new LoginController(); // Instantiate the class
-
-router.post('/', (req, res) => loginController.login(req, res));
-router.post('/register', (req, res) => loginController.register(req, res));
 router.post('/firstlogin', (req, res) => loginController.firstLogin(req, res));
 router.post('/checkStatus', (req, res) => loginController.checkUserStatus(req, res));
-
-module.exports = router;
